requirements.txt
Flask==3.1.3
pandas==3.0.1
numpy==2.4.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3
PyPDF2==3.0.1
python-docx==1.2.0
requests==2.32.5